[Updated] Sequences from FASTA.py
Added new ability for the program to find the most common pattern in your sequence aka kmer analysis. 
Fixed bug that printed the reverse compliment strand with the fasta header
made it so that you could do multiple analyses on DNA (still updating the program for RNA Sequences
Added the ability to export your DNA Translation to the same file or different file. (just put the name of the file when asked. type the directory of your file without the name of the file. (The name and directory are joined together by the program.)