Create fna_parse.py
User guided sequence that parses a .fna or .fasta file according to how the user would like it analyzed. This program is inspired by the dna_to_fasta.py program which guided the user through a series of inputs to perform different analysis of dna files. This program is meant to parse genome .fna or .fasta files and allow the user to select specific chromosomes for analysis. 

This program was tested on the 
genome of the plasmodium falciparum to select specific chromosomes for analysis and export them to a seperate .fna file. This code is part 
of a much larger 'pipeline' to search for specific sequences in a genome which I am currently working on.