Part of gene search program
Here is some of the code I have worked on in development of my gene search program to search eukaryotic genomes for specific gene sequences. This project is to help me become more familiar with the biopython package and making my own funtions to sort through genetic information. The program currently only works on prokaryotic genomes or genes with one exon in eukaryotic genomes. This python document uses the genome of the Plasmodium Falciparum or malaria parasite which is a relatively small genome with 14 chromosomes, making it a good genome to test against. The .fna document for this genome is included in the repository.